+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## main
+## [0.1.0-rc.1] - 2020-10-19
 
 * refactor - rename `areEqual` to `equals` to be consistent with dart [#93](https://github.com/LoveCommunity/love.dart/issues/93)
 * refactor - rename `EventInterceptor` to `InterceptorWithContext` [#96](https://github.com/LoveCommunity/love.dart/issues/96)